:truck: Rename menu item type class
